+++ blender-qmm/mats/EnergyConservationGroupNode-2023.py
@@ -11,9 +11,10 @@ class EnergyConservationGroup(bpy.types.Operator):
 
     def execute(self, context):
         # DOES THE MATERIAL ALREADY EXIST?
-        ng_spec = bpy.data.node_groups.get("% Specular")
-        ng_smap = bpy.data.node_groups.get("% Specular Mapping")
-        ng_sine = bpy.data.node_groups.get("% Specular Sine Mapping")
+        ng_spec = bpy.data.node_groups.get("Specular")
+        ng_smap = bpy.data.node_groups.get("Specular Mapping")
+        ng_sine = bpy.data.node_groups.get("Specular Sine Mapping")
+        ng_lin = bpy.data.node_groups.get("Linear Mapping")
         ng_ec = bpy.data.node_groups.get("Energy Conservation")
 
         if not ng_spec:
@@ -25,6 +26,9 @@ class EnergyConservationGroup(bpy.types.Operator):
         if not ng_sine:
             self.make_sine_group()
 
+        if not ng_lin:
+            self.make_lin_group()
+
         if not ng_ec:
             self.make_ec_group()
 
@@ -32,7 +36,7 @@ class EnergyConservationGroup(bpy.types.Operator):
 
     def make_spec_group(self):
         # spec_group
-        spec_group = bpy.data.node_groups.new('% Specular', 'ShaderNodeTree')
+        spec_group = bpy.data.node_groups.new('Specular', 'ShaderNodeTree')
 
         # groupinput
         group_in = self.make_node(spec_group, 'NodeGroupInput', -1000, 0)
@@ -75,9 +79,54 @@ class EnergyConservationGroup(bpy.types.Operator):
         slinks(m_power.outputs[0], m_divide.inputs[0])
         slinks(m_divide.outputs[0], group_out.inputs[1])
 
+    def make_lin_group(self):
+        # linear_mapping_group
+        lin_group = bpy.data.node_groups.new('Linear Mapping', 'ShaderNodeTree')
+
+        # groupinput
+        group_in = self.make_node(lin_group, 'NodeGroupInput', -800, 0)
+        lin_group.inputs.new('NodeSocketColor', 'Color')
+
+        # groupoutput
+        group_out = self.make_node(lin_group, 'NodeGroupOutput', 0, 0)
+        lin_group.outputs.new('NodeSocketColor', 'Color')
+
+        # combineecolor
+        lin_combine_color = self.make_node(lin_group, 'ShaderNodeCombineColor', 200, 0)
+
+        # mathmultiply
+        m_multiply = self.make_math_node(lin_group, 'MULTIPLY', -400, 200)
+        m_multiply.inputs[0].default_value = 2
+        m_multiply.use_clamp = True
+
+        # mathmultiply2
+        m_multiply2 = self.make_math_node(lin_group, 'MULTIPLY', -400, 0)
+        m_multiply2.inputs[0].default_value = 2
+        m_multiply2.use_clamp = True
+
+        # mathmultiply3
+        m_multiply3 = self.make_math_node(lin_group, 'MULTIPLY', -400, -200)
+        m_multiply3.inputs[0].default_value = 2
+        m_multiply3.use_clamp = True
+
+        # separatecolor
+        lin_separate_color = self.make_node(lin_group, 'ShaderNodeSeparateColor', -600, 0)
+        
+        lin_links = lin_group.links.new
+
+        # connect lm group
+        lin_links(group_in.outputs[0], lin_separate_color.inputs[0])
+        lin_links(lin_separate_color.outputs[0], m_multiply.inputs[1])
+        lin_links(lin_separate_color.outputs[0], m_multiply2.inputs[1])
+        lin_links(lin_separate_color.outputs[0], m_multiply3.inputs[1])
+        lin_links(m_multiply.outputs[0], lin_combine_color.inputs[0])
+        lin_links(m_multiply.outputs[0], lin_combine_color.inputs[1])
+        lin_links(m_multiply.outputs[0], lin_combine_color.inputs[2])
+        lin_links(lin_combine_color.outputs[0], group_out.inputs[0])
+
     def make_sm_group(self):
         # specular_sine_mapping_group
-        sm_group = bpy.data.node_groups.new('% Specular Mapping', 'ShaderNodeTree')
+        sm_group = bpy.data.node_groups.new('Specular Mapping', 'ShaderNodeTree')
 
         # groupinput
         group_in = self.make_node(sm_group, 'NodeGroupInput', -1000, 0)
@@ -118,7 +167,7 @@ class EnergyConservationGroup(bpy.types.Operator):
 
     def make_sine_group(self):
         # sine_mapping_group
-        sine_group = bpy.data.node_groups.new('% Specular Sine Mapping', 'ShaderNodeTree')
+        sine_group = bpy.data.node_groups.new('Specular Sine Mapping', 'ShaderNodeTree')
 
         # groupinput
         group_in = self.make_node(sine_group, 'NodeGroupInput', -800, 0)
@@ -131,14 +180,14 @@ class EnergyConservationGroup(bpy.types.Operator):
         # combineecolor
         combine_color = self.make_node(sine_group, 'ShaderNodeCombineColor', 200, 0)
 
-        # % specular mapping
-        sm_group_r = self.make_node_group(sine_group, "% Specular Mapping", '% Specular Mapping')
+        # specular mapping
+        sm_group_r = self.make_node_group(sine_group, "Specular Mapping", 'Specular Mapping')
         sm_group_r.location = (-400, 200)
 
-        sm_group_g = self.make_node_group(sine_group, "% Specular Mapping", '% Specular Mapping')
+        sm_group_g = self.make_node_group(sine_group, "Specular Mapping", 'Specular Mapping')
         sm_group_g.location = (-400, 0)
 
-        sm_group_b = self.make_node_group(sine_group, "% Specular Mapping", '% Specular Mapping')
+        sm_group_b = self.make_node_group(sine_group, "Specular Mapping", 'Specular Mapping')
         sm_group_b.location = (-400, -200)
 
         # separatecolor
@@ -166,7 +215,8 @@ class EnergyConservationGroup(bpy.types.Operator):
         ec_group.inputs.new('NodeSocketColor', 'Diffuse (Base)')    #1
         ec_group.inputs.new('NodeSocketString', '- Specular -')     #2
         ec_group.inputs.new('NodeSocketColor', 'Specular Custom')   #3
-        ec_group.inputs.new('NodeSocketFloat', 'Auto/Custom')       #4
+        ec_group.inputs.new('NodeSocketFloat', 'Custom/Auto')       #4
+        ec_group.inputs.new('NodeSocketFloat', 'Metal/Dialectric')  #5
         ec_group.inputs[0].default_value = 1.52
         ec_group.inputs[0].min_value = 0
         ec_group.inputs[0].max_value = 10
@@ -176,6 +226,9 @@ class EnergyConservationGroup(bpy.types.Operator):
         ec_group.inputs[4].default_value = 0
         ec_group.inputs[4].min_value = 0
         ec_group.inputs[4].max_value = 1
+        ec_group.inputs[5].default_value = 0
+        ec_group.inputs[5].min_value = 0
+        ec_group.inputs[5].max_value = 1
 
         # groupoutput
         group_out = self.make_node(ec_group, 'NodeGroupOutput', 0, 0)
@@ -185,52 +238,71 @@ class EnergyConservationGroup(bpy.types.Operator):
         ec_group.outputs.new('NodeSocketFloat', 'IOR')
 
         # IOR TO SPECULAR
-        spec_group = self.make_node_group(ec_group, "% Specular", '% Specular')
-        spec_group.location = (-400, -100)
+        spec_group = self.make_node_group(ec_group, "Specular", 'Specular')
+        spec_group.location = (-400, -200)
 
         # CLEARCOAT
         # mathmultiply
-        m_multiply = self.make_math_node(ec_group, 'MULTIPLY', -200, 100)
+        m_multiply = self.make_math_node(ec_group, 'MULTIPLY', -200, -200)
         m_multiply.inputs[1].default_value = 10
         m_multiply.label = "Clearcoat"
 
         # colormix
         if bv < (3, 4, 0):
-            m_colormix = self.make_node(ec_group, 'ShaderNodeMixRGB', -400, 300)
+            m_colormix = self.make_node(ec_group, 'ShaderNodeMixRGB', -600, 200)
         else:
-            m_colormix = self.make_node(ec_group, 'ShaderNodeMix', -400, 200)
+            m_colormix = self.make_node(ec_group, 'ShaderNodeMix', -600, 200)
             m_colormix.data_type = 'RGBA'
 
         # FRESNEL COLOR
         # fresnel
-        m_fresnel = self.make_node(ec_group, 'ShaderNodeFresnel', -600, 200)
-
-        # groupinput2
-        group_in2 = self.make_node(ec_group, 'NodeGroupInput', -800, 200)
+        m_fresnel = self.make_node(ec_group, 'ShaderNodeFresnel', -800, 200)
 
         # colormix2
         if bv < (3, 4, 0):
-            m_colormix2 = self.make_node(ec_group, 'ShaderNodeMixRGB', -1000, 200)
+            m_colormix2 = self.make_node(ec_group, 'ShaderNodeMixRGB', -800, -100)
         else:
-            m_colormix2 = self.make_node(ec_group, 'ShaderNodeMix', -1000, 200)
+            m_colormix2 = self.make_node(ec_group, 'ShaderNodeMix', -800, -100)
             m_colormix2.data_type = 'RGBA'
 
-        # Specular Sine Mapping
-        sine_group = self.make_node_group(ec_group, "% Specular Sine Mapping", '% Specular Sine Mapping')
-        sine_group.location = (-1200, 200)
+        # groupinput2
+        group_in2 = self.make_node(ec_group, 'NodeGroupInput', -1000, 100)
 
         # mathgreaterthan
-        m_greaterthan = self.make_math_node(ec_group, 'GREATER_THAN', -1200, 0)
+        m_greaterthan = self.make_math_node(ec_group, 'GREATER_THAN', -1000, -100)
         m_greaterthan.inputs[1].default_value = 0.5
 
+        # colormix3
+        if bv < (3, 4, 0):
+            m_colormix3 = self.make_node(ec_group, 'ShaderNodeMixRGB', -1000, -300)
+        else:
+            m_colormix3 = self.make_node(ec_group, 'ShaderNodeMix', -1000, -300)
+            m_colormix3.data_type = 'RGBA'
+
+        # mathgreaterthan2
+        m_greaterthan2 = self.make_math_node(ec_group, 'GREATER_THAN', -1200, -300)
+        m_greaterthan2.inputs[1].default_value = 0.5
+
+        # HSV
+        m_hsv = self.make_node(ec_group, 'ShaderNodeHueSaturation', -1200, -500)
+        m_hsv.inputs[2].default_value = 0.01
+
         # groupinput3
-        group_in3 = self.make_node(ec_group, 'NodeGroupInput', -1400, 200)
+        group_in3 = self.make_node(ec_group, 'NodeGroupInput', -1400, -200)
+
+        # Linear Mapping
+        lin_group = self.make_node_group(ec_group, "Linear Mapping", 'Linear Mapping')
+        lin_group.location = (-1400, -500)
+
+        # groupinput4
+        group_in4 = self.make_node(ec_group, 'NodeGroupInput', -1600, -500)
 
         links = ec_group.links.new
 
         # connect ior calc
-        links(group_in.outputs[0], spec_group.inputs[0])
         links(spec_group.outputs[0], group_out.inputs[1])
+        links(group_in.outputs[0], group_out.inputs[3])
+        links(group_in.outputs[0], spec_group.inputs[0])
 
         # connect clearcoat
         links(spec_group.outputs[0], m_multiply.inputs[0])
@@ -240,22 +312,37 @@ class EnergyConservationGroup(bpy.types.Operator):
 
         # connect color mix
         links(m_fresnel.outputs[0], m_colormix.inputs[0])
-        links(m_greaterthan.outputs[0], m_colormix2.inputs[0])
         if bv < (3, 4, 0):
             links(group_in2.outputs[1], m_colormix.inputs[1])
             links(m_colormix2.outputs[0], m_colormix.inputs[2])
             links(m_colormix.outputs[0], group_out.inputs[0])
-            links(sine_group.outputs[0], m_colormix2.inputs[1])
-            links(group_in3.outputs[3], m_colormix2.inputs[2])
         else:
             links(group_in2.outputs[1], m_colormix.inputs[6])
             links(m_colormix2.outputs[2], m_colormix.inputs[7])
             links(m_colormix.outputs[2], group_out.inputs[0])
-            links(sine_group.outputs[0], m_colormix2.inputs[6])
-            links(group_in3.outputs[3], m_colormix2.inputs[7])
 
-        links(group_in3.outputs[1], sine_group.inputs[0])
+        # connect color mix 2
+        links(m_greaterthan.outputs[0], m_colormix2.inputs[0])
+        if bv < (3, 4, 0):
+            links(group_in3.outputs[3], m_colormix2.inputs[1])
+            links(m_colormix3.outputs[0], m_colormix2.inputs[2])
+        else:
+            links(group_in3.outputs[3], m_colormix2.inputs[6])
+            links(m_colormix3.outputs[2], m_colormix2.inputs[7])
+
+        # connect color mix 3
+        links(m_greaterthan2.outputs[0], m_colormix3.inputs[0])
+        if bv < (3, 4, 0):
+            links(lin_group.outputs[0], m_colormix3.inputs[1])
+            links(m_hsv.outputs[0], m_colormix3.inputs[2])
+        else:
+            links(lin_group.outputs[0], m_colormix3.inputs[6])
+            links(m_hsv.outputs[0], m_colormix3.inputs[7])
+
+        links(lin_group.outputs[0], m_hsv.inputs[4])
         links(group_in3.outputs[4], m_greaterthan.inputs[0])
+        links(group_in3.outputs[5], m_greaterthan2.inputs[0])
+        links(group_in4.outputs[1], lin_group.inputs[0])
 
     def make_node(self, group, arg1, arg2, arg3):
         result = group.nodes.new(arg1)